rewrite: totally removed the TypedNavigator class from the source code.
