+++ src/Hl7.Fhir.Core/ElementModel/PocoBuilderExtensions.cs
@@ -30,15 +30,6 @@ namespace Hl7.Fhir.ElementModel
         public static T ToPoco<T>(this ITypedElement element, PocoBuilderSettings settings = null) where T : Base =>
                (T)element.ToPoco(settings);
 
-#pragma warning disable 612, 618
-        public static Base ToPoco(this IElementNavigator navigator, Type pocoType = null,
-            PocoBuilderSettings settings = null) => navigator.ToSourceNode().ToPoco(pocoType, settings);
-
-        public static T ToPoco<T>(this IElementNavigator navigator, PocoBuilderSettings settings = null) where T : Base =>
-               (T)navigator.ToPoco(typeof(T), settings);
-#pragma warning restore 612, 618
-
-
         public static Model.Quantity ParseQuantity(this ITypedElement instance)
         {
             var newQuantity = new Quantity
@@ -148,35 +139,5 @@ namespace Hl7.Fhir.ElementModel
         }
 
         public static string GetString(this IEnumerable<ITypedElement> instance) => instance.SingleOrDefault()?.Value as string;
-
-        [Obsolete("Use ParseCodeableConcept(this ITypedElement instance) instead")]
-        public static CodeableConcept ParseCodeableConcept(this IElementNavigator instance)
-            => ParseCodeableConcept(instance.ToTypedElement());
-
-
-        [Obsolete("Use GetString(this IEnumerable<ITypedElement> instance) instead")]
-        public static string GetString(this IEnumerable<IElementNavigator> instance)
-            => instance.SingleOrDefault()?.Value as string;
-
-
-        [Obsolete("Use ParseQuantity(this ITypedElement instance) instead")]
-        public static Model.Quantity ParseQuantity(this IElementNavigator instance) => ParseQuantity(instance.ToTypedElement());
-
-        [Obsolete("Use ParseBindable(this ITypedElement instance) instead")]
-        public static Element ParseBindable(this IElementNavigator instance) => ParseBindable(instance.ToTypedElement());
-
-
-        [Obsolete("Use ParsePrimitive<T>(this ITypedElement instance) instead")]
-        public static T ParsePrimitive<T>(this IElementNavigator instance) where T : PrimitiveType, new()
-            => ParsePrimitive<T>(instance.ToTypedElement());
-
-        [Obsolete("Use ParseCoding(this ITypedElement instance) instead")]
-        public static Coding ParseCoding(this IElementNavigator instance)
-            => ParseCoding(instance.ToTypedElement());
-
-        [Obsolete("Use ParseResourceReference(this ITypedElement instance) instead")]
-        public static ResourceReference ParseResourceReference(this IElementNavigator instance)
-             => ParseResourceReference(instance.ToTypedElement());
-
     }
 }
