+++ src/Hl7.Fhir.Specification/Validation/Validator.cs
@@ -109,38 +109,6 @@ namespace Hl7.Fhir.Validation
             return Validate(instance, declaredTypeProfile: null, statedCanonicals: null, statedProfiles: structureDefinitions).RemoveDuplicateMessages(); ;
         }
 
-        #region Obsolete public methods
-        [Obsolete("Use Validate(ITypedElement instance) instead")]
-        public OperationOutcome Validate(IElementNavigator instance)
-        {
-            return Validate(instance.ToTypedElement(), declaredTypeProfile: null, statedCanonicals: null, statedProfiles: null).RemoveDuplicateMessages(); ;
-        }
-
-        [Obsolete("Use Validate(ITypedElement instance, params string[] definitionUris) instead")]
-        public OperationOutcome Validate(IElementNavigator instance, params string[] definitionUris)
-        {
-            return Validate(instance.ToTypedElement(), (IEnumerable<string>)definitionUris).RemoveDuplicateMessages(); ;
-        }
-
-        [Obsolete("Use Validate(ITypedElement instance, IEnumerable<string> definitionUris) instead")]
-        public OperationOutcome Validate(IElementNavigator instance, IEnumerable<string> definitionUris)
-        {
-            return Validate(instance.ToTypedElement(), declaredTypeProfile: null, statedCanonicals: definitionUris, statedProfiles: null).RemoveDuplicateMessages(); 
-        }
-
-        [Obsolete("Use Validate(ITypedElement instance, params StructureDefinition[] structureDefinitions) instead")]
-        public OperationOutcome Validate(IElementNavigator instance, params StructureDefinition[] structureDefinitions)
-        {
-            return Validate(instance.ToTypedElement(), (IEnumerable<StructureDefinition>)structureDefinitions).RemoveDuplicateMessages(); 
-        }
-
-        [Obsolete("Use Validate(ITypedElement instance, IEnumerable<StructureDefinition> structureDefinitions) instead")]
-        public OperationOutcome Validate(IElementNavigator instance, IEnumerable<StructureDefinition> structureDefinitions)
-        {
-            return Validate(instance.ToTypedElement(), declaredTypeProfile: null, statedCanonicals: null, statedProfiles: structureDefinitions).RemoveDuplicateMessages(); 
-        }
-        #endregion
-
         // This is the one and only main entry point for all external validation calls (i.e. invoked by the user of the API)
         internal OperationOutcome Validate(ITypedElement instance, string declaredTypeProfile, IEnumerable<string> statedCanonicals, IEnumerable<StructureDefinition> statedProfiles)
         {
